requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
numpy==1.20.2
opencv-python-headless==4.5.1.48
Pillow==8.2.0
pyTelegramBotAPI==3.7.7
requests==2.25.1
urllib3==1.26.4
wincertstore==0.2
